perf(component): optimize tree and upload components
