+++ src/components/Tree/index.ts
@@ -2,4 +2,4 @@ import BasicTree from './src/Tree.vue';
 
 export { BasicTree };
 export type { ContextMenuItem } from '/@/hooks/web/useContextMenu';
-export * from './src/types';
+export * from './src/typing';
