+++ src/components/Tree/src/TreeHeader.vue
@@ -43,7 +43,14 @@
   import { useI18n } from '/@/hooks/web/useI18n';
   import { useDebounceFn } from '@vueuse/core';
 
-  import { ToolbarEnum } from './enum';
+  enum ToolbarEnum {
+    SELECT_ALL,
+    UN_SELECT_ALL,
+    EXPAND_ALL,
+    UN_EXPAND_ALL,
+    CHECK_STRICTLY,
+    CHECK_UN_STRICTLY,
+  }
 
   interface MenuInfo {
     key: ToolbarEnum;
