--- src/components/Tree/src/enum.ts
@@ -1,8 +1,0 @@
-export enum ToolbarEnum {
-  SELECT_ALL,
-  UN_SELECT_ALL,
-  EXPAND_ALL,
-  UN_EXPAND_ALL,
-  CHECK_STRICTLY,
-  CHECK_UN_STRICTLY,
-}
