+++ src/components/Upload/index.ts
@@ -1,1 +1,4 @@
-export { default as BasicUpload } from './src/BasicUpload.vue';
+import { withInstall } from '/@/utils';
+import basicUpload from './src/BasicUpload.vue';
+
+export const BasicUpload = withInstall(basicUpload);
