--- src/components/Upload/src/FileList.less
@@ -1,31 +1,0 @@
-.file-table {
-  width: 100%;
-  border-collapse: collapse;
-
-  .center {
-    text-align: center;
-  }
-
-  .left {
-    text-align: left;
-  }
-
-  .right {
-    text-align: right;
-  }
-
-  &-th,
-  &-td {
-    padding: 12px 8px;
-  }
-
-  thead {
-    background-color: @background-color-light;
-  }
-
-  table,
-  td,
-  th {
-    border: 1px solid @border-color-base;
-  }
-}
