+++ src/components/Upload/src/data.tsx
@@ -1,15 +1,14 @@
 import type { BasicColumn, ActionItem } from '/@/components/Table';
-
-import { FileItem, PreviewFileItem, UploadResultStatus } from './types';
+import { FileItem, PreviewFileItem, UploadResultStatus } from './typing';
 import {
   // checkImgType,
   isImgTypeByName,
 } from './helper';
 import { Progress, Tag } from 'ant-design-vue';
-
 import TableAction from '/@/components/Table/src/components/TableAction.vue';
 import ThumbUrl from './ThumbUrl.vue';
 import { useI18n } from '/@/hooks/web/useI18n';
+
 const { t } = useI18n();
 
 // 文件上传列表
